cloud-run/app/requirements.txt
flask==3.1.0
google-cloud-bigquery==3.35.1
google-api-python-client==2.169.0
gunicorn==23.0.0
